requirements.txt
alembic>=1.13.1
databases==0.9.0
fastapi>=0.111.0
pydantic>=2.7.2
PyMySQL==1.1.1
SQLAlchemy==2.0.30
mariadb==1.1.10
aiomysql==0.2.0
uvicorn>=0.31.0
opentelemetry-instrumentation-openai==0.42.0
opentelemetry-sdk                   ==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-instrumentation-pymysql==0.56b0
opentelemetry-instrumentation-fastapi==0.56b0
